Summary
This Project is for an NE555 timer circuit created in Altium. The NE555 comonent was created, added to the circuit schematic, and an accompanying library was automatically generated. The BOM for the project was then exported and included in the project files.

Whenever I attempted to link to component to digikey, Altium would crash. After googling around for a solution yielded no results, I gave up in hopes of asking someone more knowlegable at the next meeting :)